pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Slavyan <GlockOglock@yandex.ru>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
validators = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
